Updated services page
