--- pages/services/_slug.vue
@@ -1,302 +1,0 @@
-<template>
-  <div id="work" :class="service.class + ' wrapper'">
-    <section id="intro">
-      <div class="container">
-        <div class="row">
-          <div class="col-xs-12 text-center logo wow pulse">
-            <img :src="getImageUrl(service.logo)" :alt="service.title" />
-          </div>
-        </div>
-      </div>
-    </section>
-    <section id="info">
-      <div class="container">
-        <div class="row">
-          <div class="col-sm-6 wow fadeInLeft">
-            <h2>{{ service.subtitle }}</h2>
-            <hr />
-            <p>{{ service.description }}</p>
-            <template v-if="service.url != ''">
-              <br />
-              <p>
-                <a :href="service.url" class="btn btn-default" target="_blank"
-                  ><i class="fa fa-binoculars"></i> Ver en vivo</a
-                >
-              </p>
-            </template>
-          </div>
-        </div>
-      </div>
-    </section>
-    <template v-if="service.texts != ''">
-      <section id="texts">
-        <div class="container">
-          <div class="row">
-            <div
-              v-for="(text, index) in service.texts"
-              :key="index"
-              class="col-sm-8"
-              style="color:black"
-            >
-              <h2>{{ text.title }}</h2>
-              <p>{{ text.body }}</p>
-            </div>
-          </div>
-        </div>
-      </section>
-    </template>
-    <div
-      v-for="(work, index) in todos"
-      :key="index"
-      :class="[work.category] + ' col-xs-12 col-sm-12 col-lg-4'"
-    >
-      <nuxt-link
-        :to="
-          localePath({
-            name: 'projects-slug',
-            params: { slug: work.url }
-          })
-        "
-      >
-        <div
-          :key="work.name"
-          class="work"
-          :class="
-            work.favourite === true ? 'highlighted ' + work.class : work.class
-          "
-        >
-          <div class="work-header">
-            <!-- <b-img
-              v-if="work.logo != ''"
-              :src="getImageUrlAUX(work.logo)"
-              :alt="work.title"
-            /> -->
-            <!-- <h4 v-else>{{ work.title }}</h4> -->
-          </div>
-          <div class="work-body">
-            <div class="work-title">
-              {{ work.title }}
-            </div>
-          </div>
-        </div>
-      </nuxt-link>
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  data() {
-    return {
-      slug: this.$route.params.slug,
-      services: [
-        {
-          id: 'participaciondigital',
-          title: 'Participación Digital - CONSUL',
-          subtitle: 'Participación Digital - CONSUL',
-          class: 'participaciondigital',
-          description:
-            'CONSUL es la mejor plataforma de participación existente actualmente. Su potencia y versatilidad la ha llevado a ser usada por los Ayuntamientos de las principales capitales de España y Latinoamérica y recibir el Premio de Servicio Público 2018 de ONU. ENREDA instala, adapta, configura y desarrolla CONSUL a las necesidades de tu organización.',
-          logo: 'logo_consul.png',
-          url: 'http://participacionyciudadania.com/',
-          images: [
-            {
-              url: 'works/screenshot-oviedo.png'
-            },
-            {
-              url: 'works/screenshot-oviedo.png'
-            },
-            {
-              url: 'works/screenshot-oviedo.png'
-            }
-          ],
-          texts: [
-            {
-              title: 'Libre',
-              body:
-                'Al ser una plataforma de software libre todo su código está a disposición de cualquier individuo y entidad.'
-            },
-            {
-              title: 'Adaptable',
-              body:
-                'Cualquier entidad o administración puede modificarlo y utilizarlo libremente, adaptándolo a sus objetivos y necesidades.'
-            },
-            {
-              title: 'Seguro',
-              body:
-                'CONSUL cuenta con un sistema de registro que garantiza la privacidad y la neutralidad de la participación. Además puede integrarse con sistemas de padrón o censo existentes.'
-            },
-            {
-              title: 'Apoyo Continuado',
-              body:
-                'Por parte del equipo técnico y operativo encargado de su desarrollo principal, así como de las demás instituciones que impulsan el proyecto en todo el mundo.'
-            }
-          ]
-        },
-        {
-          id: 'participacionciudadana',
-          title: 'Participación Ciudadana',
-          subtitle: 'Participación Ciudadana',
-          class: 'participacionciudadana',
-          description:
-            'ENREDA posee 10 años de experiencia desarrollado procesos participativos. Estos son algunos de los servicios que podemos ofrecerte.',
-          logo: 'works/camion.png',
-          url: 'http://participacionyciudadania.com/',
-          images: [
-            {
-              url: 'works/camion.png'
-            },
-            {
-              url: 'works/camion.png'
-            },
-            {
-              url: 'works/camion.png'
-            }
-          ],
-          texts: [
-            {
-              title: 'Diseño y dinamización de presupuestos participativos',
-              body:
-                'Las metodologías participativas al servicio de la democracia presupuestaria.'
-            },
-            {
-              title: 'Escuela de ciudadanía',
-              body:
-                'Capacitación y formación para mejorar la calidad democrática en tu ciudad.'
-            },
-            {
-              title:
-                'Formación y asesoramiento en herramientas de participación online',
-              body:
-                'ENREDA es partner de la herramienta CONSUL usada por muchos ayuntamientos como plataforma de participación ciudadana. Podemos asesorar en su implatanción, formación a técnicos y asistencia en procesos participativos.'
-            },
-            {
-              title: 'Reglamentos de Participación',
-              body:
-                'Consultoría para la redacción de reglamentos o su adaptación a la Ley Andaluza de Participación.'
-            },
-            {
-              title: 'Diseño de procesos',
-              body:
-                'Diseño e implementación de planes de participación: juvenil, planeamiento urbanístico, fomento del asociacionismo., itervención social...'
-            }
-          ]
-        },
-        {
-          id: 'disenodesarrolloproducto',
-          title: 'Diseño y desarrollo de productos',
-          subtitle: 'Diseño y desarrollo de productos',
-          class: 'disenodesarrolloproducto',
-          description: '',
-          logo: 'services/diseno_desarrollo.jpg',
-          url: '',
-          images: [
-            {
-              url: ''
-            }
-          ],
-          texts: [
-            {
-              title: '',
-              body:
-                'ENREDA diseña, desarrolla y gestiona todo tipo de servicios y productos relacionados con el desarrollo de aplicaciones y uso de las nuevas tecnologías de la información y la comunicación. Elaboramos soluciones tecnológicas para entidades sociales, cooperativas, administraciones públicas, organizaciones ciudadanas y empresas.'
-            },
-            {
-              title: '',
-              body:
-                'En ENREDA desarrollamos sistemas desacoplados, escalables y adaptados a las necesidades reales de cada caso, pensando siempre en reducir el coste de gestión y en optimizar el coste del beneficio intangible.'
-            }
-          ]
-        },
-        {
-          id: 'transparencia',
-          title: 'Diseño y desarrollo de productos',
-          subtitle: 'Diseño y desarrollo de productos',
-          class: 'transparencia',
-          description: '',
-          logo: 'works/screenshot-tp.png',
-          url: 'http://transparenciapresupuestaria.enreda.coop/',
-          images: [
-            {
-              url: ''
-            }
-          ],
-          texts: [
-            {
-              title: 'Visualiza su presupuesto',
-              body:
-                'Web de visualización de datos en la que se intenta hacer comprensible el presupuesto de cualquier entidad. Además se compara con años anteriores para generar referencias y contextos en los datos. Una oportunidad para empezar a explicar los números.'
-            },
-            {
-              title: '',
-              body: 'Mecanismo de comparación de datos.'
-            },
-            {
-              title: '',
-              body: 'Herramienta orientada a lo gráfico.'
-            },
-            {
-              title: '',
-              body: '100% compatible con cualquier tecnología web.'
-            },
-            {
-              title: '',
-              body: 'Compara y comprende desde lo visual.'
-            }
-          ]
-        },
-        {
-          id: 'analitica-bigdata',
-          title: 'Analitica - Big data',
-          subtitle: 'Analitica - Big data',
-          class: 'analitica-bigdata',
-          description: '',
-          logo: 'services/big_data.jpg',
-          url: 'http://lynxanalytica.com/',
-          images: [
-            {
-              url: 'services/lynx_analytica.png'
-            }
-          ],
-          texts: [
-            {
-              title: 'Lynx Analytica',
-              body:
-                'Esta herramienta de geomarketing electoral, desarrollada por Lynx Analytica, te ofrece la información y las capacidades del Big Data para microsegmentar y geolocalizar a cualquier público que busques. Millones de datos analizados y optimizados para ser usados de forma sencilla, y ayudarte a anticipar el comportamiento político de cualquier territorio. Con esta herramienta podrás diseñar una campaña más eficiente y persuasiva, priorizar los territorios donde tendrás más éxito, microsegmentar tus mensajes según los públicos, encontrar a esos públicos allí donde viven y conseguir, en definitiva, mejorar tus resultados electorales a la vez que reduces costes. Pasa de decisiones basadas en la intuición a decisiones guiadas por la información y adelántate a tus competidores.'
-            }
-          ]
-        }
-      ]
-    }
-  },
-  computed: {
-    service() {
-      return this.services.find(service => service.id === this.slug)
-    },
-    todos() {
-      const res = this.$store.state.projects.works.find(project =>
-        project.category.includes(this.slug)
-      )
-      if (typeof res === 'object') {
-        return [res]
-      } else {
-        return res
-      }
-    }
-  },
-  nuxtI18n: {
-    paths: {
-      es: '/servicios/:slug',
-      en: '/services/:slug'
-    }
-  },
-  methods: {
-    getImageUrl(imageId) {
-      return require(`~/assets/images/${imageId}`)
-    },
-    getImageUrlAUX(imageId) {
-      return require(`~/assets/images/works/${imageId}`)
-    }
-  }
-}
-</script>
