+++ pages/services/analitica-bigdata.vue
@@ -1,0 +1,19 @@
+<template>
+  <Service slug="analitica-bigdata" />
+</template>
+
+<script>
+import Service from '~/components/Service.vue'
+
+export default {
+  components: {
+    Service
+  },
+  nuxtI18n: {
+    paths: {
+      es: '/servicios/analitica-bigdata',
+      en: '/services/analitica-bigdata'
+    }
+  }
+}
+</script>
