+++ pages/services/disenodesarrolloproducto.vue
@@ -1,0 +1,19 @@
+<template>
+  <Service slug="disenodesarrolloproducto" />
+</template>
+
+<script>
+import Service from '~/components/Service.vue'
+
+export default {
+  components: {
+    Service
+  },
+  nuxtI18n: {
+    paths: {
+      es: '/servicios/disenodesarrolloproducto',
+      en: '/services/disenodesarrolloproducto'
+    }
+  }
+}
+</script>
