+++ store/data.js
@@ -1,0 +1,409 @@
+export const state = () => ({
+  services: [
+    {
+      id: 'participaciondigital',
+      title: 'Participación Digital - CONSUL',
+      subtitle: 'Participación Digital - CONSUL',
+      class: 'participaciondigital',
+      description:
+        'CONSUL es la mejor plataforma de participación existente actualmente. Su potencia y versatilidad la ha llevado a ser usada por los Ayuntamientos de las principales capitales de España y Latinoamérica y recibir el Premio de Servicio Público 2018 de ONU. ENREDA instala, adapta, configura y desarrolla CONSUL a las necesidades de tu organización.',
+      logo: 'logo_consul.png',
+      url: 'http://participacionyciudadania.com/',
+      images: [
+        {
+          url: 'works/screenshot-oviedo.png'
+        },
+        {
+          url: 'works/screenshot-oviedo.png'
+        },
+        {
+          url: 'works/screenshot-oviedo.png'
+        }
+      ],
+      texts: [
+        {
+          title: 'Libre',
+          body:
+            'Al ser una plataforma de software libre todo su código está a disposición de cualquier individuo y entidad.'
+        },
+        {
+          title: 'Adaptable',
+          body:
+            'Cualquier entidad o administración puede modificarlo y utilizarlo libremente, adaptándolo a sus objetivos y necesidades.'
+        },
+        {
+          title: 'Seguro',
+          body:
+            'CONSUL cuenta con un sistema de registro que garantiza la privacidad y la neutralidad de la participación. Además puede integrarse con sistemas de padrón o censo existentes.'
+        },
+        {
+          title: 'Apoyo Continuado',
+          body:
+            'Por parte del equipo técnico y operativo encargado de su desarrollo principal, así como de las demás instituciones que impulsan el proyecto en todo el mundo.'
+        }
+      ]
+    },
+    {
+      id: 'participacionciudadana',
+      title: 'Participación Ciudadana',
+      subtitle: 'Participación Ciudadana',
+      class: 'participacionciudadana',
+      description:
+        'ENREDA posee 10 años de experiencia desarrollado procesos participativos. Estos son algunos de los servicios que podemos ofrecerte.',
+      logo: 'works/camion.png',
+      url: 'http://participacionyciudadania.com/',
+      images: [
+        {
+          url: 'works/camion.png'
+        },
+        {
+          url: 'works/camion.png'
+        },
+        {
+          url: 'works/camion.png'
+        }
+      ],
+      texts: [
+        {
+          title: 'Diseño y dinamización de presupuestos participativos',
+          body:
+            'Las metodologías participativas al servicio de la democracia presupuestaria.'
+        },
+        {
+          title: 'Escuela de ciudadanía',
+          body:
+            'Capacitación y formación para mejorar la calidad democrática en tu ciudad.'
+        },
+        {
+          title:
+            'Formación y asesoramiento en herramientas de participación online',
+          body:
+            'ENREDA es partner de la herramienta CONSUL usada por muchos ayuntamientos como plataforma de participación ciudadana. Podemos asesorar en su implatanción, formación a técnicos y asistencia en procesos participativos.'
+        },
+        {
+          title: 'Reglamentos de Participación',
+          body:
+            'Consultoría para la redacción de reglamentos o su adaptación a la Ley Andaluza de Participación.'
+        },
+        {
+          title: 'Diseño de procesos',
+          body:
+            'Diseño e implementación de planes de participación: juvenil, planeamiento urbanístico, fomento del asociacionismo., itervención social...'
+        }
+      ]
+    },
+    {
+      id: 'disenodesarrolloproducto',
+      title: 'Diseño y desarrollo de productos',
+      subtitle: 'Diseño y desarrollo de productos',
+      class: 'disenodesarrolloproducto',
+      description: '',
+      logo: 'services/diseno_desarrollo.jpg',
+      url: '',
+      images: [
+        {
+          url: ''
+        }
+      ],
+      texts: [
+        {
+          title: '',
+          body:
+            'ENREDA diseña, desarrolla y gestiona todo tipo de servicios y productos relacionados con el desarrollo de aplicaciones y uso de las nuevas tecnologías de la información y la comunicación. Elaboramos soluciones tecnológicas para entidades sociales, cooperativas, administraciones públicas, organizaciones ciudadanas y empresas.'
+        },
+        {
+          title: '',
+          body:
+            'En ENREDA desarrollamos sistemas desacoplados, escalables y adaptados a las necesidades reales de cada caso, pensando siempre en reducir el coste de gestión y en optimizar el coste del beneficio intangible.'
+        }
+      ]
+    },
+    {
+      id: 'transparencia',
+      title: 'Diseño y desarrollo de productos',
+      subtitle: 'Diseño y desarrollo de productos',
+      class: 'transparencia',
+      description: '',
+      logo: 'works/screenshot-tp.png',
+      url: 'http://transparenciapresupuestaria.enreda.coop/',
+      images: [
+        {
+          url: ''
+        }
+      ],
+      texts: [
+        {
+          title: 'Visualiza su presupuesto',
+          body:
+            'Web de visualización de datos en la que se intenta hacer comprensible el presupuesto de cualquier entidad. Además se compara con años anteriores para generar referencias y contextos en los datos. Una oportunidad para empezar a explicar los números.'
+        },
+        {
+          title: '',
+          body: 'Mecanismo de comparación de datos.'
+        },
+        {
+          title: '',
+          body: 'Herramienta orientada a lo gráfico.'
+        },
+        {
+          title: '',
+          body: '100% compatible con cualquier tecnología web.'
+        },
+        {
+          title: '',
+          body: 'Compara y comprende desde lo visual.'
+        }
+      ]
+    },
+    {
+      id: 'analitica-bigdata',
+      title: 'Analitica - Big data',
+      subtitle: 'Analitica - Big data',
+      class: 'analitica-bigdata',
+      description: '',
+      logo: 'services/big_data.jpg',
+      url: 'http://lynxanalytica.com/',
+      images: [
+        {
+          url: 'services/lynx_analytica.png'
+        }
+      ],
+      texts: [
+        {
+          title: 'Lynx Analytica',
+          body:
+            'Esta herramienta de geomarketing electoral, desarrollada por Lynx Analytica, te ofrece la información y las capacidades del Big Data para microsegmentar y geolocalizar a cualquier público que busques. Millones de datos analizados y optimizados para ser usados de forma sencilla, y ayudarte a anticipar el comportamiento político de cualquier territorio. Con esta herramienta podrás diseñar una campaña más eficiente y persuasiva, priorizar los territorios donde tendrás más éxito, microsegmentar tus mensajes según los públicos, encontrar a esos públicos allí donde viven y conseguir, en definitiva, mejorar tus resultados electorales a la vez que reduces costes. Pasa de decisiones basadas en la intuición a decisiones guiadas por la información y adelántate a tus competidores.'
+        }
+      ]
+    }
+  ],
+  works: [
+    {
+      title: 'Consul - El Puerto de Santa María',
+      logo: 'logo-consul-elpuerto.png',
+      class: 'aytoelpuerto',
+      url: 'consul-puerto-santa-maria',
+      category: ['consul', 'participacion', 'ayuntamiento'],
+      favourite: false
+    },
+    {
+      title: 'Consul - Oviedo',
+      logo: 'logo-ayto-oviedo.png',
+      class: 'decideoviedo',
+      url: 'consul-oviedo',
+      category: ['consul', 'participacion', 'ayuntamiento'],
+      favourite: true
+    },
+    {
+      title: 'Consul - Arcos de la Frontera',
+      logo: 'logo-ayto-arcos.png',
+      class: 'decidearcos',
+      url: 'consul-arcos-frontera',
+      category: ['consul', 'participacion', 'ayuntamiento'],
+      favourite: false
+    },
+    {
+      title: 'Consul - Santiago de Compostela',
+      logo: 'logo-ayto-santiago.png',
+      class: 'decidesantiago',
+      url: 'consul-santiago',
+      category: ['consul', 'participacion', 'ayuntamiento'],
+      favourite: false
+    },
+    {
+      title: 'decide.cadiz.es',
+      logo: '',
+      class: 'decidecadiz',
+      url: 'consul-cadiz',
+      category: ['consul', 'participacion', 'ayuntamiento'],
+      favourite: false
+    },
+    {
+      title: 'TIPI Ciudadano',
+      logo: 'logo-tipi.png',
+      class: 'tipi',
+      url: 'tipi-ciudadano',
+      category: ['transparencia', 'analisisdatos'],
+      favourite: true
+    },
+    {
+      title: 'LaCentral.coop',
+      logo: 'logo-lacentralcoop.png',
+      class: 'lacentralcoop',
+      url: 'lacentralcoop',
+      category: ['desarrolloproducto', 'economiasocial'],
+      favourite: false
+    },
+    {
+      title: 'Somos.coop',
+      logo: 'logo-somoscoop.png',
+      class: 'somoscoop',
+      url: 'somoscoop',
+      category: ['economiasocial', 'desarrolloproducto'],
+      favourite: false
+    },
+    {
+      title: 'Poletika',
+      logo: 'logo-poletika.png',
+      class: 'poletika',
+      url: 'poletika',
+      category: ['transparencia'],
+      favourite: false
+    },
+    {
+      title: 'Social to Crowd',
+      logo: 'logo-stc.png',
+      class: 'socialtocrowd',
+      url: 'socialtocrowd',
+      category: ['innovacionsocial', 'desarrolloproducto'],
+      favourite: false
+    },
+    {
+      title: 'Transparencia Presupuestaria',
+      logo: 'logo-tp.png',
+      class: 'transparenciapresupuestaria',
+      url: 'transparencia-presupuestaria',
+      category: ['desarrolloproducto', 'transparencia'],
+      favourite: false
+    },
+    {
+      title: 'CLIC Cervantes',
+      logo: 'logo-cervantes.png',
+      class: 'clic',
+      url: 'clic-cervantes',
+      category: ['desarrollotecnologico'],
+      favourite: false
+    },
+    {
+      title: 'Conil Hospeda',
+      logo: 'logo-conilhospeda.svg',
+      class: 'conilhospeda',
+      url: 'conil-hospeda',
+      category: [
+        'desarrolloproducto',
+        'economiasocial',
+        'desarrollotecnologico'
+      ],
+      favourite: false
+    },
+    {
+      title: 'En la Órbita',
+      logo: 'logo-elo.png',
+      class: 'enlaorbita',
+      url: 'enlaorbita',
+      category: ['desarrolloproducto', 'desarrollotecnologico'],
+      favourite: false
+    },
+    {
+      title: 'ICAS',
+      logo: 'logo-icas.png',
+      class: 'icas',
+      url: 'icas',
+      category: ['desarrollotecnologico'],
+      favourite: false
+    },
+    {
+      title: 'Festival de Sevilla de Cine',
+      logo: 'logo-festivalcine.png',
+      class: 'festivalcine',
+      url: 'festivalcine',
+      category: ['desarrollotecnologico'],
+      favourite: false
+    },
+    {
+      title: 'Foro Andaluz de Gobiernos Locales y Procesos Participativos',
+      logo: 'logo-foro-participacion.png',
+      class: 'foroparticipacion',
+      url: 'foro-participacion/',
+      category: ['presuspuestoparticipativo', 'participacion', 'intervencion'],
+      favourite: false
+    },
+    {
+      title: 'Reglamento de Participación de la Diputación de Cádiz',
+      logo: 'logo-diputacion-cadiz.png',
+      class: 'dipuparticipacioncadiz',
+      url: 'dipu-participacion-cadiz',
+      category: ['transparencia', 'participacion', 'reglamento'],
+      favourite: false
+    },
+    {
+      title: 'Reglamento de Participación del Ayuntamiento de Cádiz',
+      logo: '',
+      class: 'aytoparticipacioncadiz',
+      url: 'ayto-participacion-cadiz',
+      category: ['transparencia', 'participacion', 'reglamento'],
+      favourite: false
+    },
+    {
+      title: 'Formación sobre la Ley Andaluza de Participación',
+      logo: '',
+      class: 'formacionleyparticipacion',
+      url: 'formacion-ley-participacion',
+      category: ['formacion', 'participacion'],
+      favourite: false
+    },
+    {
+      title: 'Sistema Público de Participación Digital de Andalucía',
+      logo: '',
+      class: 'sistemaparticipacion',
+      url: 'sistema-participacion',
+      category: ['consul', 'participacion', 'participaciondigital'],
+      favourite: true
+    },
+    {
+      title: 'Participación para la planificación urbanística',
+      logo: 'logo-ayto-rota.png',
+      class: 'ciudadamable',
+      url: 'ciudad-amable',
+      category: [
+        'participacion',
+        'ayuntamiento',
+        'intervencion',
+        'planificacion'
+      ],
+      favourite: false
+    },
+    {
+      title: 'Presupuesto Participativos de Marbella',
+      logo: 'logo_marbella.png',
+      class: 'presupuestomarbella',
+      url: 'presupuesto-marbella',
+      category: ['presuspuestoparticipativo', 'participacion', 'ayuntamiento'],
+      favourite: false
+    },
+    {
+      title: 'Presupuesto Participativos de Mijas',
+      logo: 'logo-ayto-mijas.png',
+      class: 'presupuestomijas',
+      url: 'presupuesto-mijas',
+      category: [
+        'consul',
+        'presuspuestoparticipativo',
+        'participacion',
+        'ayuntamiento'
+      ],
+      favourite: false
+    },
+    {
+      title: 'Plan de intervención',
+      logo: 'logo-ayto-coria.png',
+      class: 'planintervencion',
+      url: 'plan-intervencion',
+      category: [
+        'participacion',
+        'ayuntamiento',
+        'intervencion',
+        'planificacion'
+      ],
+      favourite: false
+    },
+    {
+      title: 'Presupuesto Participativo de Tarifa',
+      logo: 'logo-ayto-tarifa.png',
+      class: 'presupuestotarifa',
+      url: 'presupuesto-tarifa',
+      category: ['presuspuestoparticipativo', 'participacion', 'ayuntamiento'],
+      favourite: false
+    }
+  ]
+})
